Adicionado classe de conexão com MySql, e classe para dar insert, e listar produtos

+++ src/Produto.java
@@ -24,7 +24,7 @@ public class Produto {
         return quantidade;
     }
 
-    public double getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
